AddAuthorizationAdmin

+++ CafeteriaWebApi/Controllers/AdminController.cs
@@ -4,6 +4,7 @@ using CafeteriaWebApi.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace CafeteriaWebApi.Controllers
 {
@@ -26,12 +27,19 @@ namespace CafeteriaWebApi.Controllers
 
             try
             {
+                string role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role).Value.ToString();
+
                 var res = _adminService.GetAdmins();
-                if (res == null)
+                if (role == "Admin") 
                 {
-                    return BadRequest(res);
+                    
+                    if (res == null)
+                    {
+                        return BadRequest(res);
+                    }
+                    return Ok(res);
                 }
-                return Ok(res);
+                return Forbid();
 
             } catch (Exception ex)
             {
@@ -39,62 +47,30 @@ namespace CafeteriaWebApi.Controllers
             }
         }
 
-        [HttpGet("GetProducts")]
-        public IActionResult GetProducts()
-        {
-            try
-            {
-                var res = _adminService.GetProducts();
-                if (res == null)
-                {
-                    return BadRequest(res);
-                }
-                return Ok(res);
-
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, "Error interno del servidor" + ex.Message);
-            }
-        }
-
-        [HttpGet("GetClients")]
-        public IActionResult GetClients()
-        {
-
-            try
-            {
-                var res = _adminService.GetClients();
-                if (res == null)
-                {
-                    return BadRequest(res);
-                }
-                return Ok(res);
-
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, "Error interno del servidor" + ex.Message);
-            }
-        }
 
         [HttpPost("CreateAdmin")]
         public IActionResult CreateAdmin([FromBody] AdminPostDto adminDto)
         {
             try
             {
-                var res = new Admin()
+                string role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role).Value.ToString();
+                if (role == "Admin")
                 {
-                    Email = adminDto.Email,
-                    Name = adminDto.Name,
-                    Password = adminDto.Password,
-                };
-                if (res == null)
-                {
-                    return BadRequest(res);
+
+                    var res = new Admin()
+                    {
+                        Email = adminDto.Email,
+                        Name = adminDto.Name,
+                        Password = adminDto.Password,
+                    };
+                    if (res == null)
+                    {
+                        return BadRequest(res);
+                    }
+                    int id = _userService.CreateUser(res);
+                    return Ok(id);
                 }
-                int id = _userService.CreateUser(res);
-                return Ok(id);
+                return Forbid();
 
             }
             catch (Exception ex)
@@ -104,82 +80,33 @@ namespace CafeteriaWebApi.Controllers
 
         }
 
-
-        [HttpPost("CreateClient")]
-        public IActionResult CreateClient([FromBody] AdminPostDto adminDto)
-        {
-            try
-            {
-                var res = new Client()
-                {
-                    Email = adminDto.Email,
-                    Name = adminDto.Name,
-                    Password = adminDto.Password,
-                };
-                if (res == null)
-                {
-                    return BadRequest(res);
-                }
-                int id = _userService.CreateUser(res);
-                return Ok(id);
-
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, "Error interno del servidor" + ex.Message);
-            }
-
-        }
 
         [HttpPut("UpdateAdmin/{AdminId}")]
         public IActionResult UpdateAdmin([FromRoute] int AdminId, [FromBody] UserPutDto userDto)
         {
             try
             {
-                var userToUpdate = new Admin()
+                string role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role).Value.ToString();
+                if (role == "Admin")
                 {
-                    Id = AdminId,
-                    Email = userDto.Email,
-                    Name = userDto.Name,
-                    Password = userDto.Password,
-                };
+                    var userToUpdate = new Admin()
+                    {
+                        Id = AdminId,
+                        Email = userDto.Email,
+                        Name = userDto.Name,
+                        Password = userDto.Password,
+                    };
 
-                int updatedUserId = _userService.UpdateUser(userToUpdate);
+                    int updatedUserId = _userService.UpdateUser(userToUpdate);
 
-                if (updatedUserId == 0)
-                {
-                    return NotFound($"Usuario con ID {AdminId} no encontrado");
+                    if (updatedUserId == 0)
+                    {
+                        return NotFound($"Usuario con ID {AdminId} no encontrado");
+                    }
+
+                    return Ok(updatedUserId);
                 }
-
-                return Ok(updatedUserId);
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(500, "Error interno del servidor: " + ex.Message);
-            }
-        }
-
-        [HttpPut("UpdateClient/{ClientId}")]
-        public IActionResult UpdateClient([FromRoute] int ClientId, [FromBody] UserPutDto userDto)
-        {
-            try
-            {
-                var clientToUpdate = new Client()
-                {
-                    Id = ClientId,
-                    Email = userDto.Email,
-                    Name = userDto.Name,
-                    Password = userDto.Password,
-                };
-
-                int updatedClientId = _userService.UpdateUser(clientToUpdate);
-
-                if (updatedClientId == 0)
-                {
-                    return NotFound($"Usuario con ID {ClientId} no encontrado");
-                }
-
-                return Ok(updatedClientId);
+                return Forbid();
             }
             catch (Exception ex)
             {
@@ -193,12 +120,17 @@ namespace CafeteriaWebApi.Controllers
         {
             try
             {
-                if (userId == 0)
+                string role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role).Value.ToString();
+                if (role == "Admin")
                 {
-                    return NotFound();
+                    if (userId == 0)
+                    {
+                        return NotFound();
+                    }
+                    _userService.DeleteUser(userId);
+                    return NoContent();
                 }
-                _userService.DeleteUser(userId);
-                return NoContent();
+                return Forbid();
 
             }catch (Exception ex)
             {
